(AKA-57) Вынес расширения в отдельные файлы, создал файл с константами, адаптировал установку камеры над картой
