+++ Map/Scenes/Map scene/MapViewController+UITabBarControllerDelegate.swift
@@ -1,0 +1,28 @@
+//
+//  MapViewController+UITabBarControllerDelegate.swift
+//  Map
+//
+//  Created by Arkadiy Grigoryanc on 16.01.2020.
+//
+
+import UIKit
+
+extension MapViewController: UITabBarControllerDelegate
+{
+	func tabBarController(_ tabBarController: UITabBarController, shouldSelect viewController: UIViewController) -> Bool {
+		tabBarController.delegate = nil
+		guard let navigationController = tabBarController
+			.viewControllers?
+			.first(where: { $0 is UINavigationController }) as? UINavigationController,
+			let smartTargetListViewController = navigationController.topViewController as? SmartTargetListViewController else {
+				return false
+		}
+		smartTargetListViewController.router.dataStore?.didUpdateAllSmartTargets = false
+
+		if viewController === navigationController {
+			router.routeToSmartTargetList(smartTargetListViewController)
+			return false
+		}
+		return true
+	}
+}
